+++ envs/dev/views/blocks/DateTestBlock.php
@@ -1,0 +1,3 @@
+<? print_r($vars); ?>
+
+<?= date("d.m.Y H:i", $vars['datetime']); ?>