+++ modules/admin/src/resources/js/directives.js
@@ -548,12 +548,7 @@
             	$scope.pickerPreselect = new Date();
 
             	$scope.$watch(function() { return $scope.model }, function(n, o) {
-            		
-            		if (n == o) {
-            			return;
-            		}
-            		
-            		if (n != null && n !== undefined) {
+            		if (n != null && n != undefined) {
             			var datep = new Date(n*1000);
             			$scope.pickerPreselect = datep;
             			$scope.date = $filter('date')(datep, 'dd.MM.yyyy');
@@ -637,11 +632,7 @@
 
             	$scope.$watch(function() { return $scope.model }, function(n, o) {
             		
-            		if (n == o) {
-            			return;
-            		}
-            		
-            		if (n != null && n !== undefined) {
+            		if (n != null && n != undefined) {
             			var datep = new Date(n*1000);
             			$scope.pickerPreselect = datep;
             			$scope.date = $filter('date')(datep, 'dd.MM.yyyy');
@@ -1162,7 +1153,7 @@
 
                 scope.$watch(function() { return scope.ngModel }, function(n, o) {
                     if (n != 0 && n != null && n !== undefined) {
-                        var filtering = $filter('filter')(scope.filesData, {id: n}, true);
+                        var filtering = $filter('filter')(scope.filesData, {id: n});
                         if (filtering && filtering.length == 1) {
                             scope.fileinfo = filtering[0];
                         }
@@ -1200,7 +1191,7 @@
 
                 $scope.$watch('fileId', function(n, o) {
                     if (n != 0 && n != null && n !== undefined) {
-                    	var filtering = $filter('filter')($scope.filesData, {id: n}, true);
+                    	var filtering = $filter('filter')($scope.filesData, {id: n});
                         if (filtering && filtering.length == 1) {
                         	$scope.fileinfo = filtering[0];
                         }